requirements.txt
scrapy
toolkity
jinja2
custom-redis>=3.1.2
parsel